+++ lib/Search.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { classNames } from './helpers/styles';
 import useKeyDown from './hooks/use-key-down';
+import useCustomTheme from './hooks/use-custom-theme';
 import CommandIcon from './components/icons/CommandIcon';
 import SearchIcon from './components/icons/SearchIcon';
 import Key from './components/key/Key';
@@ -24,18 +25,24 @@ export type Config = {
   docSearch?: DocSearch;
 }
 
+export type Theme = {
+  primaryColor?: string;
+}
+
 export type SearchProps = {
   config: Config;
   placeholder?: string;
   size?: 'small' | 'middle' | 'large';
   shape?: 'square' | 'round';
+  theme?: Theme;
 }
 
 export default function Search ({
   config,
   placeholder = 'Search',
   size = 'middle',
-  shape = 'round'
+  shape = 'round',
+  theme
 }: SearchProps) {
   const [searchModal, setSearchModal] = useState(false);
 
@@ -47,6 +54,7 @@ export default function Search ({
     setSearchModal(false);
   }
 
+  useCustomTheme(theme);
   useKeyDown(openSearchModal);
 
   return (
@@ -75,5 +83,5 @@ export default function Search ({
         onClose={closeSearchModal}
       />
     </>
-  )
+  );
 }