+++ lib/components/icons/BackArrowIcon.tsx
@@ -8,5 +8,5 @@ export default function BackArrowIcon ({ className }: BackArrowIconProps) {
       xmlns="http://www.w3.org/2000/svg">
       <path fill="none" strokeLinecap="round" strokeLinejoin="round" strokeWidth="48" d="M244 400L100 256l144-144M120 256h292"></path>
     </svg>
-  )
+  );
 }