+++ vite.config.ts
@@ -30,4 +30,4 @@ export default defineConfig({
   ssr: {
     noExternal: ['react-markdown']
   }
-})
+});