+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.5] - 2023-04-10
+
+- 🆕 Custom theme - Set up primary color.
+- 🛠 ESLint config.
+
 ## [0.1.4] - 2023-04-07
 
 - 🆕 Improved source url card information.
@@ -13,8 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.3] - 2023-04-06
 
 - 🆕 Add and remove recent searches from local storage.
-- 🐞 Improve logo colors in dark mode.
 - 🛠 Unify export default and types exports.
+- 🐞 Improve logo colors in dark mode.
 
 ## [0.1.2] - 2023-04-06